auto-claude: subtask-1-4 - Extract Logger class to shared module (TypeScript)
- Create new src/gondolin_integration/logging.ts with shared Logger class
- Update client.ts to import Logger from logging module
- Update sandbox.ts to import Logger from logging module
- Update service.ts to import Logger from logging module
- Remove duplicate Logger class definitions from all three files

This eliminates code duplication and provides a single source of truth
for structured JSON logging across the Gondolin integration.
